% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{FEAST_select}
\alias{FEAST_select}
\title{Feature selection by FEAST}
\usage{
FEAST_select(
  count_all_notna,
  subject_all_notna,
  Ncluster,
  Nfeature = 500,
  thre = 2,
  FAST = FALSE,
  num_pcs = 10,
  dim_reduce = c("pca", "svd", "irlba"),
  split = FALSE,
  batch_size = 1000
)
}
\arguments{
\item{count_all_notna}{a count expression matrix.}

\item{subject_all_notna}{a subject ID array.}

\item{Ncluster}{the number of input clusters.}

\item{Nfeature}{the number of features selected.}

\item{thre}{the threshold of minimum number of cells expressing a certain gene.}

\item{FAST}{boolean. If T, using FEAST_fast() for feature selection. For extreme large dataset (sample size >5000), FEAST_fast() will be automatically applied.}

\item{num_pcs}{the number of top pcs that will be investigated through the consensus clustering embedded in FEAST.}

\item{dim_reduce}{dimension reduction methods chosen from pca, svd, or irlba.}

\item{split}{boolean. If T, using subsampling to calculate the gene-level significance.}

\item{batch_size}{when split is true, need to claim the batch size for spliting the cells.}
}
\value{
A list containing the processed count matrix, subject ID array, the index of filtered cells, and the rankings of the gene-significance.
}
\description{
This function serves for feature selection based on FEAST
}
\examples{

data<- FEAST_select(count, subject, Ncluster = 6)
data <- FEAST_select(count, subject, Ncluster = 6, dim_reduce = "irlba")
data <- FEAST_select(count, subject, Ncluster = 6, Nfeature = 1000, FAST = TRUE)


}
