% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{FitPolya}
\alias{FitPolya}
\title{Run EDClust for single-cell RNA data clustering}
\usage{
FitPolya(
  count_all_notna,
  subject_all_notna,
  alpha_0,
  EM_steps = 100L,
  MM_steps = 3L,
  BaseID = 0,
  flag = 1e-04
)
}
\arguments{
\item{count_all_notna}{a count expression matrix without missing value.}

\item{subject_all_notna}{a subject ID array without missing value.}

\item{alpha_0}{the initial value of cell type effect based on InitVal_S() or InitVal()}

\item{EM_steps}{the maximum number of EM iterations. By default, EM_steps = 100, i.e.,
EDClust() will stop and return clustering results when EM_steps = 100.}

\item{MM_steps}{the maximum number of MM iterations within the M-step in each EM iteration. By default, MM_steps = 3.}

\item{BaseID}{the baseline ID used in generating alpha_0. If there's only one sample, you can set BaseID as 0.}

\item{flag}{the stopping_criterion of EDClust. By default, flag = 3.}
}
\value{
EDClust returns a list object containing:

\itemize{
  \item mem: an array of clustering label
  \item loglik: the final log likelihood after iterations
  \item alpha_0: a matrix of cell-type specific effects estimates
  \item delta: a three dimensions matrix of subject-type specific effects estimates
  \item alpha: a three dimensions matrix of overall effects estimates
  \item p: a matrix with probability that each cell belongs to each cluster
}
}
\description{
EDClust: An EM-MM hybrid method for cell clustering in multiple-subject single-cell RNA sequencing
}
\examples{

EMMM_Result <- FitPolya(count_all_notna, subject_all_notna, alpha_0)

EMMM_Result <- FitPolya(count_all_notna, subject_all_notna, alpha_0,
                        EM_steps = 20L, MM_steps = 5L, BaseID = 1, flag = 1e-05)


}
