% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{InitVal}
\alias{InitVal}
\title{Compute initial value}
\usage{
InitVal(Y, celltype, subjectid)
}
\arguments{
\item{Y}{a count expression matrix without missing value.}

\item{celltype}{a cell label array, which could be based on other clustering methods or prior biological knowledge.}

\item{subjectid}{a subject ID array.}
}
\value{
A list containing the initial value of cell-type effect based on each subject.
}
\description{
Compute initial value
}
\examples{

SLabel <- SHARP(data, N.cluster = Ncluster, rN.seed = 1234)

alpha_0 <- InitVal(data, SLabel$pred_clusters, subjectid)$alpha[[baseline_ID]]


}
