% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{InitVal_L}
\alias{InitVal_L}
\title{Run Liger for initial value}
\usage{
InitVal_L(count_all_notna, subject_all_notna, Ncluster, ID = 1, seed = 1234)
}
\arguments{
\item{count_all_notna}{a count expression matrix.}

\item{subject_all_notna}{a subject ID array.}

\item{Ncluster}{number of clusters for the final clustering results.}

\item{ID}{baseline subject ID number.}

\item{seed}{a number used for setting seeds for SHARP to obtain reproducible results.}
}
\value{
An array containing the initial value of cell-type effect based on baseline subject.
}
\description{
Run Liger for initial value
}
\examples{

alpha_0 <- InitVal_L(count_all_notna, subject_all_notna, Ncluster = 5)

alpha_0 <- InitVal_L(count_all_notna, subject_all_notna, Ncluster = 5, ID = 3, seed = 2345)


}
