% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{InitVal_S}
\alias{InitVal_S}
\title{Run SHARP for initial value}
\usage{
InitVal_S(count_all_notna, subject_all_notna, Ncluster = NULL, ID = 1)
}
\arguments{
\item{count_all_notna}{a count expression matrix.}

\item{subject_all_notna}{a subject ID array.}

\item{Ncluster}{number of clusters for the final clustering results.}

\item{ID}{baseline subject ID number.}
}
\value{
An array containing the initial value of cell-type effect based on baseline subject.
}
\description{
Run SHARP for initial value
}
\examples{

alpha_0 <- InitVal_S(count_all_notna, subject_all_notna)

alpha_0 <- InitVal_S(count_all_notna, subject_all_notna, Ncluster = 5, ID = 3)


}
