% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{Run_Liger}
\alias{Run_Liger}
\title{Run Liger for initial cluster}
\usage{
Run_Liger(countlist, Ncluster, seed = 1234)
}
\arguments{
\item{countlist}{a list containing a count expression matrix.}

\item{Ncluster}{number of clusters for the final clustering results.}

\item{seed}{a number used for setting seeds for SHARP to obtain reproducible results.}
}
\value{
An array contains the initial cluster based on Liger, which is used for InitVal_L().
}
\description{
Run Liger for initial cluster
}
\examples{

liger_clusters <- Run_Liger(count_all_notna, Ncluster = 5)

liger_clusters <- Run_Liger(count_all_notna, Ncluster = 5, seed = 2345)


}
