% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_dependency.R
\name{setup_julia}
\alias{setup_julia}
\title{Set up Julia}
\usage{
setup_julia(path = NA, Update = FALSE)
}
\arguments{
\item{path}{path of Julia}

\item{Update}{specify whether to update EDClust's Julia package. It is recommended to set it as TRUE when EDClust is updated.}
}
\value{
The Julia interface, which is an environment with the necessary methods provided by JuliaCall
}
\description{
This function initializes Julia and the EDClust.jl package.
The first time could be long due to precompilation.
Additionally, this function will call Julia and install the required packages
if they are missing.
}
\examples{

julia <- julia_setup()

julia <- julia_setup(path = "your Julia path")


}
