% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{Baseline_select}
\alias{Baseline_select}
\title{Baseline subject selection}
\usage{
Baseline_select(count_all_notna, subject_all_notna, Ncluster)
}
\arguments{
\item{count_all_notna}{a count expression matrix.}

\item{subject_all_notna}{a subject ID array.}

\item{Ncluster}{the number of input clusters.}
}
\value{
A list containing the ID of best potential baseline and each subject' score.
The one with highest score will be selected as the baseline subject.
}
\description{
This function serves for selecting the baseline subject
}
\examples{

baseID <- Baseline_select(count, subject, Ncluster = 6)


}
