% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_dependency.R
\name{setup_julia}
\alias{setup_julia}
\title{Set up Julia}
\usage{
setup_julia(path = NA)
}
\arguments{
\item{path}{path of Julia}
}
\value{
The Julia interface, which is an environment with the necessary methods provided by JuliaCall
}
\description{
This function initializes Julia and the EDClust.jl package.
The first time could be long due to precompilation.
Additionally, this function will call Julia and install the required packages
if they are missing.
}
\examples{

julia <- julia_setup()

julia <- julia_setup(path = "your Julia path")


}
